backend/requirements.txt
flask>=2.3.0
openpyxl>=3.1.0
psycopg2-binary>=2.9.0
gunicorn>=21.0.0
